Expose GraphQL with tests (#801)
These changes extend what was done previously to allow us to query the graphql client with tests via `request`.

One thing I've had to do is set the `skipLibCheck` flag in TypeScript, as there appears to be some issues with TypeScript trying to type check node_modules. Not sure what the norm here is (being new to TS) but without this I was getting compile errors.