[V3] backups on the server side (3 of 3) (#797)
FWIW I dislike that I added this into core and am counting down the days (maybe weeks? eh... months probably) until we can yank this out and put it in a plugin. I put it in core for now so that we can keep parity a little easier with V2 until we can rework `reactotron-redux` and `reactotron-mst` to use the new plugin system.