[V3] plugin system (2 of 3) (#796)
* A start of what could be the plugin system

* Changing how the plugin is setup to allow for both a client and server plugin that works together

* change this to an any so TS doesn't yell when we try and use anything from it.

* [V3] backups on the server side (3 of 3) (#797)

FWIW I dislike that I added this into core and am counting down the days (maybe weeks? eh... months probably) until we can yank this out and put it in a plugin. I put it in core for now so that we can keep parity a little easier with V2 until we can rework `reactotron-redux` and `reactotron-mst` to use the new plugin system.