🐛 TypeScript Fixtravaganza (#812)
A few issues fixed here in the `typescript` theme...

### GraphQL

When i brought in apollo, it upped the graphql version to 14.  This was incompatible with `graphql` and `@types/graphql` 13 which was being used by `type-graphql`.

Fortunately, @rmevans9 is a wizard and showed me how to fix via `yarn resolutions`.  There's currently an [yarn issue](https://github.com/yarnpkg/yarn/issues/5039) that requires me to put the `resolutions` in both `package.json` and `packages/reactotron-server/package.json`, but no big deal.

### tslint

I brought in `ts-lint` for `reactotron-server`.

### typescript

I bumped the version to `3.0.3` cuz there's some footguns in prior versions.

### tsconfig out dir changes

When you `tsc` a folder that has 2 directories, both directories are copied to the `outDir`.  1 directory will only copy the contents (strangely enough).  So I updated the `outDir` to reflect that.

Sorry about that.  That caught me off guard.

### .gitignore

I promoted the `.cache` to the root `.gitignore`.  This is for `parcel`, but having 1 `.gitignore` is preferable since `vscode` doesn't seem to support unioning them together nicely (to make the folders dim in the explorer view).