Fix `trackGlobalError` plugin (#1292)
This PR fixes the `trackGlobalError` plugin, which appears to have been
broken for a while.

We rely on an internal utility called `parseErrorStack` that is provided
by the `react-native` package. However, the signature got changed
internally and this broke our usage:
https://github.com/facebook/react-native/commit/9edfc43aad2f68be8a37c916779681cd9ad13fee

Additionally, the `DevTools` typescript definitions file is still
incorrect. This PR adds some manual type definitions of internals that
are not provided by react-native, as well as some corrections of
internal types.

This PR also removes the "swizzled" strategy for spying on internals in
favor of using `Proxy`.

Related Issues:
- https://github.com/infinitered/reactotron/issues/1260
- https://github.com/infinitered/reactotron/issues/1253

Video of new usage:


https://github.com/infinitered/reactotron/assets/37849890/6fccedf5-df46-4391-aa16-8cb5351cd741