feat(reactotron): Add Android `adb` command interface to the Help page. (#1320)
Sometimes connecting to physical android devices can trip up people new
to reactotron or react native development.

This PR adds a section to the Help page where users can see which
android devices that are connected to their computer via USB and
automatically run common functions on each device including:

1. `adb reverse tcp:9090 tcp:9090` for reactotron connections on android
devices
2. `adb reverse tcp:8081 tcp:8081` for metro connections on android
devices
3. `adb shell input text '"RR"` to reload an active react native
application
4. `adb shell input keyevent 82` to shake the device and display the
react native developer menu. This is particularly useful for simulator
instances

...all with a simple to use interface on the Help page. This works with
multiple connected devices and will only run on the selected device.

There's also a hidden "port settings" for the tcp reverse command that
you can access via the "gear" settings icon where power users can change
the port that reactotron and metro are running on.

### To Do:

- [x] Test that this works on Windows instances
- [ ] Fail gracefully if `adb` isn't set up correctly
- [ ] Possible: feedback to the user that something actually happened
when they click "Reverse Tunnel"
- [ ] Determine a better place for this, maybe? Should this go on the
Help page?

### How It works:

When the app boots up, it defines the actions that can be taken with
`adb` as well as spawns a process for `adb track-devices` so that we can
be notified whenever an android device is connected or disconnected from
USB.

When the Help screen loads, it fires off a command to get the list of
connected devices and parses the result from the command's `stdout`. If
the `adb track-devices` process determines that a device has been
connected or disconnected, it requests the list of devices again.

Overall, this works REALLY well!

### Here's the functionality in action:

> I am physically plugging and unplugging my android device:


![667EE0DF-9FE5-41F4-BA0A-21C23C4F2091-9010-000064220203A491](https://github.com/infinitered/reactotron/assets/139261/76d0eedb-97b6-4a90-b783-3fd1340cfe30)

Here's what the interface looks like with multiple android devices
connected:

![Screenshot 2023-11-01 at 5 41
10 PM](https://github.com/infinitered/reactotron/assets/139261/618abd58-cadc-4ea3-ba75-127a44bc8d4b)

Here's the input fields where power users can change the reactotron and
metro ports:

![Screenshot 2023-11-01 at 5 46
20 PM](https://github.com/infinitered/reactotron/assets/139261/41fbb707-e5ae-47cb-a017-36c7f880fa11)