fix: BigInts on hermes android were crashing during serialization for reactotron
I though I had fixed this but we got more reports from the community that it was still broken on android.

I added a few new buttons to the example app and was able to reproduce the crash on my android emulator on expo 50 / rn 0.73.6.

With these changes, bigints show up properly in the reactotron app.