feat(reactotron-app): show redux diffs in timeline (#1531 by @camilossantos2809)
## Please verify the following:

- [x] `yarn build-and-test:local` passes
- [x] I have added tests for any new features, if relevant
- [ ] `README.md` (or relevant documentation) has been updated with your
changes

## Describe your PR



This change allows developers to see what has been modified in the store
after an action is dispatched.

It is particularly useful when an action updates multiple modules. When
an action is called in Redux, the current store is compared with the new
store, and the differences are sent to Reactotron to be displayed in the
timeline.

To determine these differences, the microdiff package was used and added
as a dependency to reactotron-redux.

I am not entirely familiar with your build process, but I assume that
the microdiff package will be installed automatically via npm/yarn when
developers upgrade the Reactotron plugin in their projects.

I'm open to feedback and suggestions for further improvements to this
implementation. Please feel free to reach out if you have any questions
or if there are areas you'd like me to refine.

closes #496 

![Screenshot 2025-01-20 at 14 12
29](https://github.com/user-attachments/assets/1528678c-1143-49ab-917f-8be9dd1a62ae)

---------

Co-authored-by: Jamon Holmgren <code@jamon.dev>
Co-authored-by: Ítalo Masserano <italo_masserano@outlook.com>
Co-authored-by: Infinite Red <ci@infinite.red>
Co-authored-by: Kedar Sarmalkar <110056501+ksarmalkarbb@users.noreply.github.com>